+++ packages/browser-crawler/src/internals/browser-crawler.ts
@@ -252,6 +252,12 @@ export interface BrowserCrawlerOptions<
      * Can be also set via {@apilink Configuration}.
      */
     headless?: boolean | 'new' | 'old'; // `new`/`old` are for puppeteer only
+
+    /**
+     * Whether to ignore custom elements (and their #shadow-roots) when processing the page content via `parseWithCheerio` helper.
+     * By default, they are expanded automatically. Use this option to disable this behavior.
+     */
+    ignoreShadowRoots?: boolean;
 }
 
 /**
@@ -494,25 +500,22 @@ export abstract class BrowserCrawler<
         const experimentalContainers = this.launchContext?.experimentalContainers;
 
         if (this.proxyConfiguration) {
-            if (useIncognitoPages || experimentalContainers) {
-                const { session } = crawlingContext;
+            const { session } = crawlingContext;
 
-                const proxyInfo = await this.proxyConfiguration.newProxyInfo(session?.id, { request: crawlingContext.request });
-                crawlingContext.proxyInfo = proxyInfo;
+            const proxyInfo = await this.proxyConfiguration.newProxyInfo(session?.id, { request: crawlingContext.request });
+            crawlingContext.proxyInfo = proxyInfo;
 
-                newPageOptions.proxyUrl = proxyInfo.url;
+            newPageOptions.proxyUrl = proxyInfo?.url;
+            newPageOptions.proxyTier = proxyInfo?.proxyTier;
 
-                if (this.proxyConfiguration.isManInTheMiddle) {
-                    /**
+            if (this.proxyConfiguration.isManInTheMiddle) {
+                /**
                      * @see https://playwright.dev/docs/api/class-browser/#browser-new-context
                      * @see https://github.com/puppeteer/puppeteer/blob/main/docs/api.md
                      */
-                    newPageOptions.pageOptions = {
-                        ignoreHTTPSErrors: true,
-                    };
-                }
-            } else {
-                newPageOptions.proxyTier = this.proxyConfiguration.getProxyTier(crawlingContext.request);
+                newPageOptions.pageOptions = {
+                    ignoreHTTPSErrors: true,
+                };
             }
         }
 
@@ -706,12 +709,12 @@ export abstract class BrowserCrawler<
             launchContextExtends.session = await this.sessionPool.getSession();
         }
 
-        if (this.proxyConfiguration) {
+        if (this.proxyConfiguration && !launchContext.proxyUrl) {
             const proxyInfo = await this.proxyConfiguration.newProxyInfo(
                 launchContextExtends.session?.id,
                 { proxyTier: (launchContext.proxyTier as number) ?? undefined },
             );
-            launchContext.proxyUrl = proxyInfo.url;
+            launchContext.proxyUrl = proxyInfo?.url;
             launchContextExtends.proxyInfo = proxyInfo;
 
             // Disable SSL verification for MITM proxies