+++ website/versioned_docs/version-3.9/guides/proxy_management_session_cheerio.ts
@@ -1,0 +1,10 @@
+import { CheerioCrawler, ProxyConfiguration } from 'crawlee';
+
+const proxyConfiguration = new ProxyConfiguration({ /* opts */ });
+
+const crawler = new CheerioCrawler({
+    useSessionPool: true,
+    persistCookiesPerSession: true,
+    proxyConfiguration,
+    // ...
+});